DRY-up node_links
