+++ api/base/views.py
@@ -23,7 +23,8 @@ from api.base.serializers import LinkedNodesRelationshipSerializer
 from api.base import utils
 from api.nodes.permissions import ReadOnlyIfRegistration
 from api.nodes.permissions import ContributorOrPublicForRelationshipPointers
-from api.base.utils import is_bulk_request
+from api.base.utils import is_bulk_request, get_user_auth
+
 
 CACHE = weakref.WeakKeyDictionary()
 
@@ -646,3 +647,27 @@ class BaseContributorList(JSONAPIBaseView, generics.ListAPIView, ListFilterMixin
                     raise ValidationError('Contributor identifier incorrectly formatted.')
             queryset[:] = [contrib for contrib in queryset if contrib._id in contrib_ids]
         return queryset
+
+class BaseNodeLinksDetail(JSONAPIBaseView, generics.RetrieveAPIView):
+
+    def get_queryset(self):
+        auth = get_user_auth(self.request)
+        return sorted([
+            pointer.node for pointer in
+            self.get_node().nodes_pointer
+            if not pointer.node.is_deleted and not pointer.node.is_collection and
+            pointer.node.can_view(auth) and not pointer.node.is_retracted
+        ], key=lambda n: n.date_modified, reverse=True)
+
+
+class BaseNodeLinksList(JSONAPIBaseView, generics.ListAPIView):
+
+    def get_queryset(self):
+        auth = get_user_auth(self.request)
+        return sorted([
+            pointer.node for pointer in
+            self.get_node().nodes_pointer
+            if not pointer.node.is_deleted and not pointer.node.is_collection and
+            pointer.node.can_view(auth) and not pointer.node.is_retracted
+        ], key=lambda n: n.date_modified, reverse=True)
+
